Added DevMailTransport.
Enable in AdditionalConfiguration using

```
$GLOBALS['TYPO3_CONF_VARS']['MAIL']['transport'] = \Vierwd\VierwdBase\Mail\DevMailTransport::class;
```

This MailTransport will send all mails to developer and not to real receivers. The From-header of all mails will
be changed to clients@4wdmedia.de. The receiver can be configured using `$TYPO3_CONF_VARS['MAIL']['defaultReceiverAddress']`.
If that is not set, the server variable `$_SERVER['VIERWD_EMAIL']` is checked and if that fails, the sendmail_path
is checked, if it contains an email address.